Removed CDW support from MedicationOrderController.
