+++ data-query-tests/src/test/resources/cdw/datamart-index.yaml
@@ -284,42 +284,6 @@ entries:
     query: /MedicationDispense:1.02?patient=43000199&type=FF,UD
     page: 1
     count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103i1200389904206O.xml
-    query: /MedicationOrder:1.03?identifier=1400162277477:O
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103i5555555555555O.xml
-    query: /MedicationOrder:1.03?identifier=5555555555555:O
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg0.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 0
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg10000ct20.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 10000
-    count: 20
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg1ct0.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 1
-    count: 0
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg1ct2.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 1
-    count: 2
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg1ct21.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 1
-    count: 21
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg2ct2.xml
-    query: /MedicationOrder:1.03?patient=43000199
-    page: 2
-    count: 2
   - file: ../../../../target/cdw/samples/MedSta102i1000000398707.xml
     query: /MedicationStatement:1.02?identifier=1000000398707
     page: 1
@@ -834,42 +798,4 @@ entries:
     query: /DiagnosticReport:1.02?patient=43000199&category=LAB
     page: 1
     count: 20
-    # Alias the 1.03 MedicationOrder samples as 1.04 to make sure they are tested.
-    # The schema was unchanged from 1.03 and new schemas and new samples were not generated.
-  - file: ../../../../target/cdw/samples/MedOrd103i1200389904206O.xml
-    query: /MedicationOrder:1.04?identifier=1400162277477:O
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103i5555555555555O.xml
-    query: /MedicationOrder:1.04?identifier=5555555555555:O
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg0.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 0
-    count: 15
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg10000ct20.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 10000
-    count: 20
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg1ct0.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 1
-    count: 0
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg1ct2.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 1
-    count: 2
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg1ct21.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 1
-    count: 20
-  - file: ../../../../target/cdw/samples/MedOrd103p185601V825290pg2ct2.xml
-    query: /MedicationOrder:1.04?patient=43000199
-    page: 2
-    count: 2
