--- data-query/src/test/resources/cdw/old-medicationorder-1.04.json
@@ -1,46 +1,0 @@
-{
-  "resourceType" : "MedicationOrder",
-  "id" : "3e768d87-b1bb-5b96-a206-81d1e0e7dd33",
-  "dateWritten" : "2015-04-15T17:22:03Z",
-  "status" : "completed",
-  "patient" : {
-    "reference" : "https://www.freedomstream.io/CDCArgonaut/api/Patient/185601V825290",
-    "display" : "VETERAN,JOHN Q"
-  },
-  "prescriber" : {
-    "reference" : "https://www.freedomstream.io/CDCArgonaut/api/Practitioner/93e4e3c3-8d8c-5b53-996f-6047d0232231",
-    "display" : "SMITH,ATTENDING D"
-  },
-  "medicationReference" : {
-    "reference" : "https://www.freedomstream.io/CDCArgonaut/api/Medication/2f773f73-ad7f-56ca-891e-8e364c913fe0",
-    "display" : "ATORVASTATIN CALCIUM 80MG TAB"
-  },
-  "dosageInstruction" : [
-    {
-      "text" : " TAKE ONE TABLET BY MOUTH ONE TIME EACH DAY FOR CHOLESTEROL",
-      "timing" : {
-        "code" : {
-          "text" : "QDAILY"
-        }
-      },
-      "route" : {
-        "text" : "ORAL"
-      },
-      "doseQuantity" : {
-        "value" : 1
-      },
-      "asNeededBoolean" : false
-    }
-  ],
-  "dispenseRequest" : {
-    "quantity" : {
-      "value" : 30
-    },
-    "expectedSupplyDuration" : {
-      "value" : 30,
-      "unit" : "days",
-      "system" : "http://unitsofmeasure",
-      "code" : "d"
-    }
-  }
-}
